+++ src/ionic-tree-provider.ts
@@ -25,7 +25,7 @@ interface IonicState {
   selectedAndroidDeviceName?: string;
   selectedIOSDeviceName?: string;
   projectDirty?: boolean; // Was there a likely change in the project (ie file saved)
-  outputIsFocused: boolean;
+  outputIsFocused: boolean; // Whether to override focus on the Output channel
   refreshDebugDevices: boolean; // Should we refresh the list of debuggable devices
   packageManager: PackageManager;
 }